Extract message manager
