+++ console_bot/console_bot.py
@@ -1,116 +1,13 @@
 """Module providing a function printing bot messages."""
 
-from .input_manager import InputManager
-from .address_book import AddressBook
-from .edit import edit_record
-from rich.console import Console
-from rich.table import Table
-from rich import box
-from rich.align import Align
-from rich.live import Live
 from prompt_toolkit import PromptSession
 from prompt_toolkit.auto_suggest import AutoSuggestFromHistory
 from prompt_toolkit.history import InMemoryHistory
 from prompt_toolkit.styles import Style
-import time
-from contextlib import contextmanager
-
-
-console = Console()
-BEAT_TIME = 0.04
-
-
-@contextmanager
-def beat(length: int = 1) -> None:
-    yield
-    time.sleep(length * BEAT_TIME)
-
-
-def print_help():
-    table = Table(title="Available commands:")
-    table.box = box.SIMPLE_HEAD
-    table.border_style = "bright_yellow"
-    table.header_style = "bold bright_blue"
-    table_centered = Align.left(table)
-    console.clear()
-    with Live(table_centered, console=console, screen=False, refresh_per_second=20):
-        with beat(10):
-            table.add_column("Name", style="magenta")
-            table.add_column("Description", style="green")
-        with beat(10):
-            table.add_row("hello", ":wave: Ask the bot how it can help you.")
-            table.add_row("help", "Shows this help message.")
-            table.add_row("close/exit", "Exits the program.")
-
-        with beat(10):
-            table.add_row(
-                "add \[name] \[phone]",
-                ":telephone_receiver: Adds a contact with the specified name and phone number. Email and address can be added using separate commands.",
-            )
-            table.add_row(
-                "change \[name] \[old-phone] \[phone]",
-                ":telephone_receiver: Changes the phone number for the specified contact.",
-            )
-            table.add_row(
-                "phone \[name]",
-                ":telephone_receiver: Retrieves the phone number for the specified contact.",
-            )
-            table.add_row(
-                "search \[term]",
-                ":magnifying_glass_tilted_left: Global search, retrives any matches in any contact's fields.",
-            )
-        with beat(10):
-            table.add_row(
-                "add-email \[name] \[email]",
-                ":e-mail: Adds an email to the specified contact.",
-            )
-            table.add_row(
-                "change-email \[name] \[old-email] \[email]",
-                ":e-mail: Changes the email for the specified contact.",
-            )
-            table.add_row(
-                "email \[name]",
-                ":e-mail: Retrieves the email for the specified contact.",
-            )
-            table.add_row(
-                "add-address \[name] \[address]",
-                ":house_with_garden: Adds an address to the specified contact.",
-            )
-            table.add_row(
-                "change-address \[name] \[old-address] \[address]",
-                ":house_with_garden: Changes the address for the specified contact.",
-            )
-            table.add_row(
-                "address \[name]",
-                ":house_with_garden: Retrieves the address for the specified contact.",
-            )
-            table.add_row("edit \[fieldname] \[name]", "Edit/Delete phone, email, address.")
-            table.add_row("all", "Displays all contacts in the system.")
-            table.add_row("random-book", "Generate random book with 10 contacts.")
-        with beat(10):
-            table.add_row(
-                "add-birthday \[name] \[birthday]", "Adds birthday to the contact."
-            )
-            table.add_row(
-                "show-birthday \[name]", "Shows birthday for specific contact."
-            )
-            table.add_row(
-                "birthdays", "Shows birthdays for all contacts celebrating next week."
-            )
-        with beat(10):
-            table.add_row("add-note", ":spiral_notepad: Adds note to Note Book.")
-            table.add_row("find-notes", ":spiral_notepad: Searches notes by keywords.")
-            table.add_row(
-                "delete-note", ":spiral_notepad: Deletes note by index in Note Book."
-            )
-            table.add_row(
-                "change-note", ":spiral_notepad: Changes note by index in Note Book."
-            )
-            table.add_row("all-notes", ":spiral_notepad: Shows all notes in Note Book.")
-            table.add_row(
-                "random-note",
-                ":spiral_notepad: Generates random note from Taras Hryhorovych Shevchenko poem",
-            )
+from .input_manager import InputManager
+from .address_book import AddressBook
+from .edit import edit_record
+from .message_manager import print_help_message, print_welcome_message
 
 commands = [
     "hello",
@@ -154,7 +51,7 @@ def main():
         auto_suggest=AutoSuggestFromHistory(),
         enable_history_search=True,
     )
-    print("Welcome to the assistant bot!")
+    print_welcome_message("Welcome to the assistant bot!")
     print(AddressBook().check_today_birthdays())
     print_note(input_manager.random_note(save=False))
 
@@ -196,7 +93,7 @@ def main():
         elif command == "all":
             print(input_manager.get_all_contacts())
         elif command == "help":
-            print_help()
+            print_help_message()
         elif command == "add-birthday":
             print(input_manager.add_birthday(args))
         elif command == "show-birthday":
